[ALPS02349318] Audio: Fix potentially overflow operation
[Detail]
function:Auddrv_CheckInterruptTiming
Unintentional integer overflow
[Solution]
Add type casting before arithmetic compuation

MTK-Commit-Id: f134b041762452fa516390aab0aac0f1cc8a7e96

CR-Id: ALPS02349318
Feature: Others

+++ sound/soc/mediatek/common/auddrv_underflow_mach.c
@@ -173,7 +173,7 @@ void Auddrv_CheckInterruptTiming(void)
 			     __func__, Irq_time_t2, Irq_time_t1, Irq_time_t1 - Irq_time_t2,
 			     mDL1_Interrupt_Interval_Limit);
 			Irq_time_t2 = Irq_time_t1 - Irq_time_t2;
-			if (Irq_time_t2 > mDL1_Interrupt_Interval_Limit * 1000000) {
+			if (Irq_time_t2 > (unsigned long long)mDL1_Interrupt_Interval_Limit * 1000000) {
 				pr_debug
 				    ("%s interrupt may be blocked Irq_time_t2 = %llu Interval_Limit = %d\n",
 				     __func__, Irq_time_t2, mDL1_Interrupt_Interval_Limit);
